modify tests to separate tests to botclient methods froms tests to chatbot methods
