+++ bin/main.rb
@@ -1,3 +1,3 @@
-require_relative 'botclient'
+require_relative '../lib/botclient'
 
 Botclient.new('1109360723:AAHam4xsAf-7wgF8Hjt6ACbxxOH66cimbaM', 'a71219e79a6b01978ac3a9f3ffccca37', true)
