+++ lib/chatbot.rb
