+++ lib/option.rb
